requirements.txt
playwright>=1.40.0
requests>=2.31.0
tenacity>=8.2.0
openpyxl>=3.1.0
pandas>=2.1.0
anthropic>=0.18.0
openai>=1.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
